Removed unused helper file. Renamed apiCalls file to cleaner and also renamed apiCalls.js to helper.js
